requirements.txt
pandas
seaborn
openpyxl
pyperformance
tox
requests